Added more hiding Spiders Added torch to the end-pit Replaced boomerang with knife in the end-chest Moved static knife to the hidden canyon Modified InventoryController.cs to deactivate after interaction, to prevent duplicate pickups Removed unused variables from PlayerController.cs and SpiderController.cs Added check for empty inventory in InteractionSpider.cs

+++ Assets/Scripts/SpiderController.cs
@@ -19,7 +19,7 @@ public class SpiderController : MonoBehaviour, IHitable, IPatrolable
     };
 
     //values that will be set in the Inspector
-    public int initialHealthMultiplier = 1;
+    public float initialHealthMultiplier = 1;
     public float initialRunSpeedMultiplier = 3f;
     public float initialWalkSpeedMultiplier = 2f;
     public int initialDamageMultiplier = 1;
